common externals path should be called last
packages for Pd-libraries are supposed to install into
/usr/lib/pd/extra if they are expected to run with all flavours of Pd,
or into /usr/lib/puredata/extra if they are specific to Pd-vanilla.
In order to allow a vanilla-specific package puredata-foo override the
more general pd-foo, /u/l/puredata should be searched before /u/l/pd

this was discussed at length with hc-steiner and r-haefeli at the pd-irc.